requirements.txt
# Core dependencies
ollama>=0.1.0              # Primary - Local LLM client
requests>=2.31.0           # For HTTP requests
python-dotenv>=1.0.0       # Environment configuration
python-dateutil>=2.8.0     # Datetime utilities
pytz>=2024.1               # Timezone support
markdown>=3.4.0            # Markdown processing
jinja2>=3.1.0              # Template rendering
pyyaml>=6.0                # YAML configuration

# Web UI
flask>=3.0.0               # Web framework
flask-cors>=4.0.0          # CORS support

# Optional (only if using cloud APIs)
# openai>=1.0.0
# anthropic>=0.18.0

